+++ src/email_user/tests/test_serializer.py
@@ -1,4 +1,7 @@
+import os
+
 import dateutil.parser
+from django.conf import settings
 
 from django.test import TestCase