+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Version history
 ===============
 
+1.0.0 (2022-07-01)
+===========================
+* Dropped support for Django 3.1 and less. Added support for Django 4.0 and Python 3.10.
+* Added Email-Case Configuration.
+
 1.0.0-alpha.2 (2021-04-15)
 ===========================
 * Updated factory